% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating_curves.R
\name{bearing_inner}
\alias{bearing_inner}
\title{Simulate change in bearing and bearing}
\usage{
bearing_inner(path, bearing_regs)
}
\arguments{
\item{path}{TC data frame, containing obs thus far in simulation}

\item{bearing_regs}{List of block-specific change in bearing regressions}
}
\value{
TC data frame, w/ simulated new bearing at current observation
}
\description{
This function updates the bearing of the TC by predicting the 
change in bearing.
}
\details{
At the current observation,  block-specific regression 
models predict change in bearing. Then normal random noise with mean 0 and
standard dev equal to the predicted standard error is added to the change in
bearing. The new bearing is determined by adding the old bearing and the
change in bearing.
}
