% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visual_functions.R
\name{bubble_data_plot_prep}
\alias{bubble_data_plot_prep}
\title{Generate Bubble CB data frames}
\usage{
bubble_data_plot_prep(bubble_step)
}
\arguments{
\item{bubble_step}{same input as \code{\link{calculateErrorBandsBubble}} 
needs}
}
\value{
\item{center}{data n x 2 points for center path of bubble}
\item{lower}{data n x 2 points for "lower" tangental point a center path}
\item{upper}{data n x 2 points for "upper" tangental point a center path}
}
\description{
Generate Bubble CB data frames
}
