% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visual_functions.R
\name{bubble_data_plot_prep_center}
\alias{bubble_data_plot_prep_center}
\title{Plot centers for bubble data}
\usage{
bubble_data_plot_prep_center(bubble_plot_data, base_graph = NULL,
  color = "pink", ...)
}
\arguments{
\item{bubble_plot_data}{list of center points, lower and upper tangential 
points from \code{\link{bubble_data_plot_prep}} function 
(each a n x 2 data frame)}

\item{base_graph}{plot to add points to}

\item{color}{color of points}
}
\value{
ggplot object with points on it
}
\description{
Plot centers for bubble data
}
