% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{calc_k_dist}
\alias{calc_k_dist}
\title{Finds the kth smallest distance for each point}
\usage{
calc_k_dist(dist_matrix, K = 7)
}
\arguments{
\item{dist_matrix}{n x k* matrix with distances (each row an observation)
k* needs to be greater than k}

\item{K}{number of neighbors desired}
}
\value{
dist_k vector with kth distance (per each row)
}
\description{
Finds the kth smallest distance for each point
}
