% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_functions.R
\name{check_points_in_bubbleCI}
\alias{check_points_in_bubbleCI}
\title{Check if points are inside bubble CI}
\usage{
check_points_in_bubbleCI(df_points, center_df, radius_df, long = 1, lat = 2)
}
\arguments{
\item{df_points}{Dataframe of points which inclusion needs to be calculated}

\item{center_df}{Dataframe of central points of the CI bubble}

\item{radius_df}{Dataframe of radius of the CI bubble}

\item{long}{column index of the longitude in the central points}

\item{lat}{column index of the latitude in the central points}
}
\value{
Boolean vectors on whether the points are in any of the bubble CIs 
or not
}
\description{
Given a dataframe of points this function returns a vector of the same
length with 1 or 0 whether the respective points is included in any of the CI
bubble.
}
