% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_functions.R
\name{check_points_within_diff_radius}
\alias{check_points_within_diff_radius}
\title{Checks if points are within at least 1 ball from the bubble CI (different 
radii)}
\usage{
check_points_within_diff_radius(tc_bubble_structure, df_points, long = 1,
  lat = 2, unit_measure = "nautical mile", verbose = TRUE)
}
\arguments{
\item{tc_bubble_structure}{Bubble CI list with centers, radius, positive 
and negative parts (at least a list of centers data frame and radius vector)}

\item{df_points}{Dataframe of points which inclusion needs to be calculated}

\item{long}{Column index of the longitude for \code{df_points} entry}

\item{lat}{Column index of the latitude for \code{df_points} entry}

\item{unit_measure}{string of measurement unit of points, see options in
datamart::uconv.}

\item{verbose}{boolean logic if should have print outs (specifically that we
switch the columns due to longlat boolean)}
}
\value{
vector of 0/1 to indicate if the query point was inside at least one 
ball
}
\description{
uses flexablity of \code{\link{get_area_diff_radius_inner}}
}
