% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_source.R
\name{convert_long}
\alias{convert_long}
\title{Convert longitude hemispheres to pos/neg}
\usage{
convert_long(df)
}
\arguments{
\item{df}{TC data frame, where long column has E/W at end}
}
\value{
Longitude column, converted to numeric
}
\description{
If longitude ends in "E", drop "E" and convert longitude to 
numeric. If longitude ends in "W", drop "W" and convert longitude to negative
numeric.
}
