% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_hull.R
\name{convex_hull_structure}
\alias{convex_hull_structure}
\title{Convex Hull creation}
\usage{
convex_hull_structure(data_list, alpha, dist_mat = NULL,
  data_deep_points = NULL, depth_vector = NULL, c_position = 1:2,
  verbose = FALSE, ...)
}
\arguments{
\item{data_list}{list of hurricanes}

\item{alpha}{for credible band (related to depth)}

\item{dist_mat}{distance matrix (otherwise is calculated)}

\item{data_deep_points}{data deep points from depth function 
(otherwise calculated)}

\item{depth_vector}{Depth vector values}

\item{c_position}{Position of the longitude/latitude pair}

\item{verbose}{if the distance matrix is verbose}

\item{...}{other parameters in distance calculation through 
\code{\link{distMatrixPath_innersq}}}
}
\value{
\item{polygon}{dataframe}
\item{size}{area of the convex hull}
\item{poly}{spatial polygon object of convex hull}
}
\description{
Convex Hull creation
}
