% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credible_interval_pipeline_functions.R
\name{credible_interval_single_tc}
\alias{credible_interval_single_tc}
\title{Extract credible intervals of 4 methods from a single test TC}
\usage{
credible_interval_single_tc(dflist, test_true_path, alpha_level, long = 1,
  lat = 2, verbose = FALSE, kde_grid_size = rep(1000, 2),
  alpha_ci_level = 0.05, unit_measure = "nautical mile")
}
\arguments{
\item{dflist}{List of simulated TCs}

\item{test_true_path}{Dataframe with the true TC path to be tested}

\item{alpha_level}{Alpha level of all the CI}

\item{long}{Column index of the longitude}

\item{lat}{Column index of the latitude}

\item{verbose}{If TRUE update messages will be displayed}

\item{kde_grid_size}{Two dimensional vector for the KDE grid size}

\item{alpha_ci_level}{Fraction for confidence interval of area estimates based on uniform
sampling distribution.}

\item{unit_measure}{Unit of measure used for distance}
}
\value{
1 list with 4 arguments, once for each of the 4 methods, which all 
have an 'area' and 'in_vec' argument to determine the area value and a 
boolean vector which reported whether the point of the true TC was inside the
TC or not.
}
\description{
This function extracts the KDE, Bubble CI, delta-ball and convex hull 
credible intervals from the simulations of a single TC. It then checks the 
coverage of such CI as number of points included in the CI for each of the 
methods
}
