% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{d2MatrixPath_pt}
\alias{d2MatrixPath_pt}
\title{Computes matrix D^2(x,y)}
\usage{
d2MatrixPath_pt(prob_matrix_t)
}
\arguments{
\item{prob_matrix_t}{(n x n) probability matrix}
}
\value{
output_mat (n x n) \eqn{D^2(x,y)} matrix
}
\description{
Computes matrix D^2(x,y) defined as follows:
   \deqn{D^2(x,y) = sum_k (P(x,k)-P(y,k))/stat_k}
Where stat_k is the stationary vector of the probability matrix.

Note: 
This function always assumes that the probability matrix is built correctly
and that all the rows sums to one, such that an eigenvalue 1 always exists - 
and it is the biggest of the series. So the function goes and looks for the 
first eigenvector of the series.

The author knows longer knows what this does.
}
