% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca_estimation_and_visualization.R
\name{data_plot_sc_paths}
\alias{data_plot_sc_paths}
\title{Compresses data from list (to be plotted in ggplot)}
\usage{
data_plot_sc_paths(test_list, scp_output, c_position = 1:2)
}
\arguments{
\item{test_list}{list of paths to analysis (each a df)}

\item{scp_output}{spectral_cluster_process function output}

\item{c_position}{positions of lat and lon columns in test_list data frames}
}
\value{
data frame that can be used to visualize curves
}
\description{
Compresses data into tidyverse focused dataframe (for ggplot)
}
\details{
\code{scp_output} is expected to be a list with at least 2 components: 
(1) \code{test_p_estimate} - probablity estimates of the test curves 
(\eqn{p_i})
(2) \code{test_weight} - a vector of normalized probabilities 
(\eqn{\p_i/\sum_k p_k})
}
