% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_function.R
\name{depth_curves_to_points}
\alias{depth_curves_to_points}
\title{Get deepest curves' points in a data frame}
\usage{
depth_curves_to_points(data_list, alpha, dist_mat = NULL, c_position = 1:2,
  depth_vector = NULL, verbose = FALSE, ...)
}
\arguments{
\item{data_list}{list of hurricanes}

\item{alpha}{for credible band (related to depth)}

\item{dist_mat}{distance matrix (otherwise is calculated)}

\item{c_position}{only needed if created 13 point reduction}

\item{depth_vector}{vector of depth values (otherwise calculated)}

\item{verbose}{if the distance matrix is verbose}

\item{...}{other parameters in distance calculation through 
`distMatrixPath_innersq`}
}
\value{
data frame with points in desired curves
}
\description{
Get deepest curves' points in a data frame
}
