% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{distAlongP}
\alias{distAlongP}
\title{Distance/bearing between points along path}
\usage{
distAlongP(data_df, output_length = "nautical mile", longlat = TRUE,
  verbose = FALSE)
}
\arguments{
\item{data_df}{(n x 2) data.frame, each row is a pair of values (lat, lon) 
unless specified the reverse with longlat boolean}

\item{output_length}{string of measurement unit of points, see options in
datamart::uconv.}

\item{longlat}{boolean if the order of the columns is longitude then latitude
- if FALSE then column order is reversed.}

\item{verbose}{boolean logic if should have print outs (specifically that we
switch the columns due to longlat boolean)}
}
\value{
\item{distance}{vector of distances beween points (n - 1)}
\item{bearing}{vector of bear change between points (n - 1)}
}
\description{
Calculates the distance and bearing between each point of a path
}
