% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{distBetweenP}
\alias{distBetweenP}
\title{Distance between points}
\usage{
distBetweenP(data_df_p1, data_df_p2, output_length = "nautical mile",
  longlat = TRUE, verbose = FALSE)
}
\arguments{
\item{data_df_p1}{(n x 2) data.frame, each row is a pair of values (lat, lon) 
unless specified the reverse with longlat boolean}

\item{data_df_p2}{(n x 2) second data.frame to be compared, each row is a 
pair of values (lat, lon) unless specified the reverse with longlat boolean}

\item{output_length}{string of measurement unit of points, see options in
datamart::uconv.}

\item{longlat}{boolean if the order of the columns is longitude then latitude
- if FALSE then column order is reversed.}

\item{verbose}{boolean logic if should have print outs (specifically that we
switch the columns due to longlat boolean)}
}
\value{
vector of n distances between each order point in the two data frames
}
\description{
Calculates the distance between each point of every path.
}
