% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{distMatrixPath_innersq}
\alias{distMatrixPath_innersq}
\title{Calculates the distance matrix Delta between n paths}
\usage{
distMatrixPath_innersq(path_mat_list, output_length = "nautical mile",
  longlat = TRUE, verbose = TRUE)
}
\arguments{
\item{path_mat_list}{list of paths (matricies/ dataframes)}

\item{output_length}{string of measurement unit of points, see options in
datamart::uconv.}

\item{longlat}{boolean if the order of the columns is longitude then latitude
- if FALSE then column order is reversed.}

\item{verbose}{boolean logic if should have print outs while computing 
distance matrix}
}
\value{
distance matrix od dimension n x n
}
\description{
Calculates the distance matrix Delta between n paths
}
