% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{distMatrixPath_t2t_path}
\alias{distMatrixPath_t2t_path}
\title{Create path distance matrix}
\usage{
distMatrixPath_t2t_path(path_mat_list_train, path_mat_list_test,
  output_length = "nautical mile", longlat = TRUE, verbose = FALSE)
}
\arguments{
\item{path_mat_list_train}{list of paths of training curves (n objects)}

\item{path_mat_list_test}{list of paths of test curves (m objects)}

\item{output_length}{string of measurement unit of points, see options in
datamart::uconv.}

\item{longlat}{boolean if the order of the columns is longitude then latitude
- if FALSE then column order is reversed.}

\item{verbose}{boolean logic if should have print outs while computing 
distance matrix}
}
\value{
delta matrix distance, output as a matrix (m x n)
}
\description{
Creates a path distance matrix between the test paths and 
training paths (t2t) contains the correct L2 norm structure sum of square 
distances
}
