% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_functions.R
\name{extract_countour}
\alias{extract_countour}
\title{Selection of specific countour level from KDE object}
\usage{
extract_countour(kde_obj, alpha_level)
}
\arguments{
\item{kde_obj}{kde object}

\item{alpha_level}{contour level which needs to be extracted. Integer from 1 
to 99.}
}
\value{
Countour at (100-level) for the kde object.
}
\description{
This function extracts a specific level of contour from a kde object.
}
\details{
Only works for some levels that the kde object calculated originally.
}
\examples{
\dontrun{
set.seed(8192)

x1 <- 2^rnorm(100)
y1 <- rnorm(100)
dfmat <- cbind(x1,y1)
kde_object <- ks::kde(dfmat)

cont <- extract_countour(kde_object, 5)
}
}
