% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_functions.R
\name{flatten_tc_list}
\alias{flatten_tc_list}
\title{Flatten TC list to single dataframe}
\usage{
flatten_tc_list(dflist)
}
\arguments{
\item{dflist}{list of TC dataframes}
}
\value{
Resulting dataframe from a row-wise addition of all TC dataframes.
}
\description{
Converts list of dataframes to single dataframes row-wise
}
\examples{
\dontrun{
set.seed(8192)
dflist <- list()
for (i in c(1:10)) {
 x <- 2^rnorm(100)
 y <- rnorm(100)
 dflist[[i]] <- cbind(x,y)
}

flatten_df <- flatten_tc_list(dflist)
dim(flatten_df)
}
}
