% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating_curves.R
\name{generate_all}
\alias{generate_all}
\title{Generate all simulated TC paths}
\usage{
generate_all(train = NA, test = NA, remove_length_2 = T,
  number_paths = 350, replicate = T, verbose = T)
}
\arguments{
\item{train}{List of train TCs on which to train simulations}

\item{test}{List of test TCs from which to simulate new paths}

\item{remove_length_2}{Boolean for whether to remove test TCs of length <= 2.
AR models need >= 3 observations to start simulation. (In train/test split
of the paper, the only TC of length 2 has been assigned to training.)}

\item{number_paths}{Number of paths to simulate for each test TC start}

\item{replicate}{Boolean for whether to replicate simulations of paper.
This reads in the train/test data from raw_data.Rdata, sets 
number_paths = 350, and sets a seed value.}

\item{verbose}{Boolean for whether to output progress}
}
\value{
List of simulated TC paths. The list length equals the number of test
TC, with one element named for each of the test TCs. Each of these elements 
is a list of length 4. These elements are a list of number_paths 
"Auto_DeathRegs" simulations, number_paths "Auto_NoDeathRegs" simulations, 
number_paths "NoAuto_DeathRegs" simulations, and number_paths 
"NoAuto_NoDeathRegs" simulations.
}
\description{
The default settings of this function generate the simulated TCs 
used in the paper. Models are trained using the training TCs. Then TC paths
are simulated from the starting paths of the test TCs.
}
