% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating_curves.R
\name{generate_curve}
\alias{generate_curve}
\title{Simulate TC path from initial observations}
\usage{
generate_curve(path, train_models, death_regs_ind, auto_ind)
}
\arguments{
\item{path}{Initial 2 or 3 (auto_ind = F or T, respectively) observations
of TC upon which to generate path}

\item{train_models}{List of trained models, generated from get_train_models}

\item{death_regs_ind}{Indicator, whether to use death regressions}

\item{auto_ind}{Indicator, whether bearing/speed regs are autoregressive}
}
\value{
Full, simulated path of TC as a column of longitudinal
coordinates and a column of latitudinal coordinates.
}
\description{
This function simulates a TC path sequentially, starting
with the initial observations.
}
\details{
At each time step, block-specific regression functions determine 
change in TC bearing/speed. Using these values, the function projects forward 
to the next point. The user specifies whether the change in bearing/speed 
functions are autoregressive or not (auto_ind = T or F). The user also
specifies whether to use block-specific death regressions or draws from overall 
kernel density to model death of TC (death_regs_ind = T or F).
}
