% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_ball.R
\name{get_area}
\alias{get_area}
\title{Estimate area of union of balls (wrapper)}
\usage{
get_area(data, delta, n = 10000, alpha = 0.05)
}
\arguments{
\item{data}{data continuous data frame of individual points (in each row)}

\item{delta}{radius of each ball inside the union}

\item{n}{number of points drawn uniformly within a box around the true data}

\item{alpha}{alpha level for 2 sided confidence interval estimate}
}
\value{
\item{area}{estimated area of union of balls}
\item{area_ci}{vector with lower and upper confidence interval estimate}
}
\description{
Wrapper function to estimate the area of union of balls with 
radius delta, using uniform draws
}
