% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_ball.R
\name{get_area_inner}
\alias{get_area_inner}
\title{Estimate area of union of balls (inner)}
\usage{
get_area_inner(data, query, size, delta, alpha = 0.05)
}
\arguments{
\item{data}{data continuous data frame of individual points (in each row)}

\item{query}{data points uniformly drawn from a space (with calculated size)}

\item{size}{size of space where query points were drawn}

\item{delta}{radius of each ball inside the union}

\item{alpha}{alpha level for 2 sided confidence interval estimate}
}
\value{
\describe{
\item{area}{estimated area of union of balls}
\item{area_ci}{vector with lower and upper confidence interval estimate}
}
}
\description{
Inner function to estimate the area of union of balls with 
radius delta, using uniform draws
}
