% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{get_bearing_speed_regs}
\alias{get_bearing_speed_regs}
\title{Determine bearing and speed regressions}
\usage{
get_bearing_speed_regs(dflist_unlist, auto)
}
\arguments{
\item{dflist_unlist}{All TC observations, including regression columns, 
as a single data frame}

\item{auto}{Boolean for whether to compute autoregressive models}
}
\value{
List of bearing_regs and speed_regs.
\item{bearing_regs}{Block-specific regressions for change in bearing}
\item{speed_regs}{Block-specific regressions for change in speed}
}
\description{
Compute block-specific regression models for change in
bearing and change in speed
}
