% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_ball.R
\name{get_box_points}
\alias{get_box_points}
\title{Sample from box around data points}
\usage{
get_box_points(data, n = 10000)
}
\arguments{
\item{data}{continuous data frame of individual points (in each row)}

\item{n}{number of points drawn}
}
\value{
\describe{
\item{box_points}{data frame of uniformly randomly drawn points}
\item{size}{the size of the box from which points are drawn}
}
}
\description{
Inner function to get uniform sampling of points in box around 
data points
}
