% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_ball.R
\name{get_delta}
\alias{get_delta}
\title{Find delta for covering}
\usage{
get_delta(data, dist_mat = NULL)
}
\arguments{
\item{data}{continuous data frame of individual points (in each row)}

\item{dist_mat}{distance matrix, calculated otherwise via euclidean distance}
}
\value{
\item{dist_mat}{distance matrix between points}
\item{mm_delta}{the minimum distance (delta)}
}
\description{
Find the minimum distance (delta) such that all points are 
within delta of at least one other point.
}
