% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{get_speed}
\alias{get_speed}
\title{Compute speed and change in speed}
\usage{
get_speed(lat, long)
}
\arguments{
\item{lat}{Latitude vector}

\item{long}{Longitude vector}
}
\value{
List of speed, speed_change, and speed_prev vectors.
\item{speed}{Estimated speed at (\code{lat[i], long[i]}). Assumes travel at this
speed from (\code{lat[i], long[i]}) to (\code{lat[i+1], long[i+1]}).}
\item{speed_change}{where \code{speed_change[i] = speed[i] - speed[i-1]}}
\item{speed_prev}{where \code{speed_prev[i] = speed[i-1]}}
}
\description{
Determine speed (meters/hr) of TC and change in speed between 
all consecutive (latitude, longitude) pairs. Also determine previous speed at
each observation.
}
\details{
The speed at observation i (in meters per hour) is the distance 
between points \eqn{i} and \eqn{i+1} (in meters) divided by 6 (hours). This 
is because there are 6 hours between consecutive observations.
}
