% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{get_starting_points}
\alias{get_starting_points}
\title{Set up initial TC starting points}
\usage{
get_starting_points(dfval, auto = T)
}
\arguments{
\item{dfval}{List of validation TC data frames}

\item{auto}{Boolean indicating whether to set up data frames for 
autoregression framework}
}
\value{
List of data frames containing initial TC observations,
as well as regression variables
}
\description{
Takes in a list of validation TC data frames. Transforms
data frames into starting observations for simulating new TCs. 
Saves starting 2/3 observations and appends regression columns.
}
