% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_ball.R
\name{get_tri_matrix}
\alias{get_tri_matrix}
\title{Make triangle matrix}
\usage{
get_tri_matrix(dtri_data_tri)
}
\arguments{
\item{dtri_data_tri}{sp object with triangles (n triangles)}
}
\value{
a matrix (n x 3) with strings of locations of 3 points in triangle
}
\description{
Makes triangle matrix for points in matrix
}
