% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca_estimation_and_visualization.R
\name{ggvis_all_weighted}
\alias{ggvis_all_weighted}
\title{Creates clean visual of weighted curves using Spectral Clustering Analysis}
\usage{
ggvis_all_weighted(scp_output, test_list, c_position = 1:2, zoom = 4,
  train_alpha = 0.3, test_color_power = 1/3, test_color_low = "white",
  test_color_high = "red")
}
\arguments{
\item{scp_output}{spectral_cluster_process function output}

\item{test_list}{list of paths to analysis (each a df)}

\item{c_position}{positions of lat and lon columns in test_list data frames}

\item{zoom}{map zoom for ggmap (plot 2: map)}

\item{train_alpha}{opacity level for black training points (plot 1: scatter)}

\item{test_color_power}{power transformation (x^ test_color_power) of 
probability values for test points color (plot 1: scatter)}

\item{test_color_low}{lower color for range of colors on test points prob
(plot 1: scatter)}

\item{test_color_high}{high color for range of colors on test points prob
(plot 1: scatter)}
}
\value{
\item{gg_path}{ggmap based map object of colored test curves}
\item{gg_proj}{ggplot scatter plot of training and colored test points}
also visualizes both graphics together using grid.arrange
}
\description{
Creates clean visual of weighted curves using Spectral Clustering Analysis.
Specifically, the projected points in 2d and the paths colored
}
\details{
ggplot experts are encouraged to use the output of this function to  
instead of just running with the created plot
}
