% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visual_functions.R
\name{ggvis_bubble_data}
\alias{ggvis_bubble_data}
\title{Plot lines for bubble data}
\usage{
ggvis_bubble_data(bubble_plot_data, base_graph = NULL, centers = FALSE,
  connect = TRUE, color = "pink", linewidth = 1, zoom = 4, ...)
}
\arguments{
\item{bubble_plot_data}{list of center points, positive and negative 
tangential points 
(each a n x 2 data frame)}

\item{base_graph}{plot to add to}

\item{centers}{boolean to decide whether to also plot centers (as points)}

\item{connect}{boolean to connect the left and right final tangential points}

\item{color}{color of band}

\item{linewidth}{width of line}

\item{zoom}{map zoom for ggmap}

\item{...}{Interior ggpath parameters}
}
\value{
ggplot object with curves on it
}
\description{
Plot lines for bubble data
}
