% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visual_functions.R
\name{ggvis_convex_hull}
\alias{ggvis_convex_hull}
\title{Visualize delta ball exterior centers (with ggplot)}
\usage{
ggvis_convex_hull(output_lines, base_graph = NULL, zoom = 4,
  color = "pink", ...)
}
\arguments{
\item{output_lines}{data frame of exterior lines}

\item{base_graph}{ggplot object for base graph 
(created from output_lines otherwise)}

\item{zoom}{map zoom for ggmap}

\item{color}{color of band}

\item{...}{interior ggpath parameters}
}
\value{
ggplot object of contour
}
\description{
Basically the same of \code{\link{ggvis_delta_ball_contour}}
}
