% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visual_functions.R
\name{ggvis_delta_ball_contour}
\alias{ggvis_delta_ball_contour}
\title{Visualize delta ball exterior centers (with ggplot)}
\usage{
ggvis_delta_ball_contour(output_lines, base_graph = NULL, zoom = 4,
  color = "pink", ...)
}
\arguments{
\item{output_lines}{data frame of exterior lines (not ordered)}

\item{base_graph}{ggplot object for base graph 
(created from output_lines otherwise)}

\item{zoom}{map zoom for ggmap}

\item{color}{color of band}

\item{...}{interior ggpath parameters}
}
\value{
ggplot object of contours
}
\description{
Visualize delta ball exterior centers (with ggplot)
}
