% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visual_functions.R
\name{ggvis_kde_contour}
\alias{ggvis_kde_contour}
\title{Create ggplot of contour}
\usage{
ggvis_kde_contour(level_contour_df, data_plot, base_graph = NULL, zoom = 4)
}
\arguments{
\item{level_contour_df}{data frame of level contour}

\item{data_plot}{data frame of points}

\item{base_graph}{ggplot object for base graph 
(created from data_out otherwise)}

\item{zoom}{map zoom for ggmap}
}
\value{
ggplot object of contour and data points.
}
\description{
Create ggplot of contour
}
