% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_map.R
\name{kernel_estimate}
\alias{kernel_estimate}
\title{KNN density estimate for each point}
\usage{
kernel_estimate(train, test, k)
}
\arguments{
\item{train}{}

\item{test}{}

\item{k}{number of neighbors that matter}
}
\value{
p density estimate for test observations
}
\description{
TODO: figure out why this isn't just knn...
}
