% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca_estimation_and_visualization.R
\name{new_points_projection}
\alias{new_points_projection}
\title{Project New Points with SCA}
\usage{
new_points_projection(P_test, psi_map_train, lambda_train)
}
\arguments{
\item{P_test}{transition matrix, \eqn{t = 1}, (n x m) calculated for the new 
points on the old points where n = number of new points (test), m is number 
of old points (train)}

\item{psi_map_train}{psi projection of training data (m x nv) where 
nv = dimension of projected space}

\item{lambda_train}{vector of lambdas of original decomposition of 
transition matrix (\eqn{t = 1}), length is nv}
}
\value{
psi_map_estimated estimated psi projection of test data
}
\description{
creates the estimate of the projection of the test values into the \eqn{\Psi}
space of training
}
