% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{pMatrixPath_average}
\alias{pMatrixPath_average}
\title{Averages P matrices}
\usage{
pMatrixPath_average(prob_matrix, t_var, t_type = "upper")
}
\arguments{
\item{prob_matrix}{(n x n) transition matrix}

\item{t_var}{integer or vector containing information about t}

\item{t_type}{string, either "upper", "vec", or "log2" 
"upper" : integer t_var, range t from 1:t_var
"vec"   : vector t_var range t along vec
"log2"  : integer t_var, range t from 2^(0-> t_var)}
}
\value{
p_average (n x n) transition matrix
}
\description{
Averages together P matrices raised to different t values

Notes: Assumes that the P matrix is well created (rows sum to 1)

The author no longer knows what this does.
}
