% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{probMatrixPath_k}
\alias{probMatrixPath_k}
\title{Create probability matrix}
\usage{
probMatrixPath_k(dist_matrix, kNN_sigma, kNN_sigma_new = kNN_sigma)
}
\arguments{
\item{dist_matrix}{n x m distance matrix}

\item{kNN_sigma}{m x 1 vector with sigma estimates for original structure}

\item{kNN_sigma_new}{n x 1 vector with sigma estimates for new structure
if dist_matrix is not symmetric, assumed new relationships are being provided}
}
\value{
p n x m matrix (transitional matrix)
}
\description{
Creates probability matrix, using KNN vector
}
