% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_source.R
\name{pull_data}
\alias{pull_data}
\title{Pull TC data from HURDAT database}
\usage{
pull_data()
}
\value{
TC data as a list of data frames
}
\description{
This function reads in data from the online HURDAT database and
converts it into a list of data frames, with one data frame per TC. 
Only data on TCs from 1951 onward are returned.
}
