% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_ball.R
\name{remove_delta_off_line}
\alias{remove_delta_off_line}
\title{Shorten line by delta on both sides}
\usage{
remove_delta_off_line(line, delta)
}
\arguments{
\item{line}{2 x 2 matrix of edge points of line}

\item{delta}{numeric delta to be subtracted}
}
\value{
2 x 2 matrix of edges that are shrunk
}
\description{
Inner function to remove delta length from both sides of a line
}
