% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_ball.R
\name{remove_duplicates_func}
\alias{remove_duplicates_func}
\title{Remove duplicate points}
\usage{
remove_duplicates_func(data_raw)
}
\arguments{
\item{data_raw}{data points of observations 
expected lat, long column names}
}
\value{
data without duplicates
}
\description{
Remove duplicates of points in point cloud (needed for certain 
algorithms)
}
