% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_ball.R
\name{remove_lines_from_tri}
\alias{remove_lines_from_tri}
\title{Remove triangles with edges that need removal}
\usage{
remove_lines_from_tri(tuples_of_tri, removed_mat)
}
\arguments{
\item{tuples_of_tri}{data frame with tuples of triangle edges and 
triangle index}

\item{removed_mat}{edges to be removed}
}
\value{
data frame with tuples of triangle not removed
}
\description{
Remove triangles from tuple matrix that have 1 or more edge that
needs to be removed

TODO: this function needs to be cleaned up
}
