% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca_estimation_and_visualization.R
\name{right_eigenvector_compression}
\alias{right_eigenvector_compression}
\title{Right Eigenvector Compression}
\usage{
right_eigenvector_compression(P, nu = 100, nv = 100, plot_n = 0, t = 5)
}
\arguments{
\item{P}{transitional matrix}

\item{nu}{number of right eigenvectors}

\item{nv}{number of left eigenvectors}

\item{plot_n}{number of original eigenvectors to plot (if any)}

\item{t}{the number of steps you want to take with the transition matrix 
(power you raise the transition matrix)}
}
\value{
\item{psi_map_x}{matrix (n x nv) of rows are \eqn{\Psi_t(x)}}
\item{lambda}{vector of original lambdas from the decomposition of P 
(length nv)}
}
\description{
Compresses P transitional matrix to approximate \eqn{D^2_t} matrix 
(diffusion distance at scale t)
}
