% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{round_curve}
\alias{round_curve}
\title{Round long/lat to one decimal point}
\usage{
round_curve(curve_df)
}
\arguments{
\item{curve_df}{Simulated TC data frame containing only a longitude column 
and a latitude column}
}
\value{
Simulated TC data frame, with rounded longitude/latitude columns
}
\description{
Rounds longitude and latitude degrees to closest tenth 
to match NOAA format
}
