% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca_estimation_and_visualization.R
\name{sca_projection}
\alias{sca_projection}
\title{Applies SCA projection of each set of simulated curves and captures the 
associated probability values in the lower dimensional space}
\usage{
sca_projection(test_data_list, train_list_13_point, train_info,
  c_position = 1:2, verbose = TRUE, curve_type_vec = c("Auto_DeathRegs",
  "Auto_NoDeathRegs", "NoAuto_DeathRegs", "NoAuto_NoDeathRegs"))
}
\arguments{
\item{test_data_list}{a list of list of paths of test curves (simulated 
curves)}

\item{train_list_13_point}{list of 13 point compression of the paths of 
training curves, each in lonlat format for the first 2 columns}

\item{train_info}{structure for SCA algorithm from training data outputted
from \code{\link{training_structure_estimating_p}}.}

\item{c_position}{position of test_data's longitude and latitude columns 
respectively}

\item{verbose}{if TRUE, progress of processing of the each set of test curves
is expressed.}

\item{curve_type_vec}{vector with the strings referring to the different 
types of curves}
}
\value{
A list information about each group of simulated curves, each element
is a list of the following components:

\item{projection}{data frame of associated projected points of test / 
simulated curves}
\item{p_estimate}{probability estimates in the lower dimensional space of the
test / simulated curves}
}
\description{
This function is a wrapper function of 
\code{\link{inner_sca_projection}}
}
