% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{shift}
\alias{shift}
\title{Shift vector forward by 1}
\usage{
shift(x)
}
\arguments{
\item{x}{Vector}
}
\value{
\code{x} vector shifted forward by 1
}
\description{
Shift vector x forward by 1, so that \code{x_shift[1] = NA,
x_shift[2] = x[1], ..., x_shift[n] = x[n-1]}.
}
