% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca_estimation_and_visualization.R
\name{spectral_cluster_process}
\alias{spectral_cluster_process}
\title{Spectral Clustering Process}
\usage{
spectral_cluster_process(test_list, train_df, D_train, K = 4, t = 1,
  dim = 5, kdensity = 10, c_position = 1:2)
}
\arguments{
\item{test_list}{list of paths to analysis (each a df)}

\item{train_df}{13 point compression of training curves}

\item{D_train}{distance matrix from training curves}

\item{K}{nearest neighbors for spectral clustering inside smoothing}

\item{t}{power to raise spectral cluster transition matrix}

\item{dim}{dimension of lower dimensional projection from spectral c.}

\item{kdensity}{nearest neighbors for density estimate in lower dim space}

\item{c_position}{nearest neighbors for density estimate in lower dim space}
}
\value{
\item{test_p_estimate}{probability estimates for test data}
\item{test_weight}{scaled probabilities estimates (by max of probs) for test}
\item{test_13point}{13 point compression for test data}
\item{test_projected}{test data in projection space}
\item{train_projected}{training data in projection space}
}
\description{
produces weights and projection for paths (from projection)

Note: This may be redudant now
}
