% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating_curves.R
\name{speed_inner}
\alias{speed_inner}
\title{Simulate change in speed and speed}
\usage{
speed_inner(path, speed_regs)
}
\arguments{
\item{path}{TC data frame, containing obs thus far in simulation}

\item{speed_regs}{List of block-specific change in speed regressions}
}
\value{
TC data frame, w/ simulated new speed at current observation
}
\description{
This function updates the speed of the TC by predicting the 
change in speed.
}
\details{
At the current observation, block-specific regression 
models predict change in speed. Then normal random noise with mean 0 and
standard dev equal to the predicted standard error is added to the change in
speed. The new speed is determined by adding the old speed and the
change in speed.
}
