% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating_curves.R
\name{update_curve_inner}
\alias{update_curve_inner}
\title{Project forward to new point}
\usage{
update_curve_inner(path)
}
\arguments{
\item{path}{TC data frame, containing obs thus far in simulation, 
as well as bearing and speed on current obs}
}
\value{
TC data frame, with simulated position of next observation
}
\description{
After simulating a new bearing and new speed, this function
projects forward to the new point.
}
\details{
This function takes path with the simulated new bearing
and simulated new speed. It projects forward to the next point, assuming
that the TC maintains that bearing and speed for the next 6 hours.
Then it fills in the regression variables on the new point.
}
