% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visual_functions.R
\name{bubble_data_plot_lines}
\alias{bubble_data_plot_lines}
\title{Plot lines for bubble data}
\usage{
bubble_data_plot_lines(bubble_plot_data, base_graph = NULL, color = "pink",
  linewidth = 1, zoom = 4, ...)
}
\arguments{
\item{bubble_plot_data}{list of center points, lower and upper tangential 
points 
(each a n x 2 data frame)}

\item{base_graph}{plot to add band to}

\item{color}{color of band}

\item{linewidth}{width of line}

\item{zoom}{map zoom for ggmap}

\item{...}{Interior ggpath parameters}
}
\value{
ggplot object with curves on it
}
\description{
Plot lines for bubble data
}
