% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca_estimation_and_visualization.R
\name{color_function}
\alias{color_function}
\title{Inner color function setup}
\usage{
color_function(weights_in, test_color_power = 1/3, test_color_low = "white",
  test_color_high = "red")
}
\arguments{
\item{weights_in}{probability weights for}

\item{test_color_power}{power to raise the probability weights 
(then use linear scaling for colors)}

\item{test_color_low}{lower color value for color range,}

\item{test_color_high}{higher color value for color range,}
}
\value{
\item{breaks}{vector of which of the 10 breaks each weight is in 
(relative to power transformation)}
\item{colors_rw}{color palette ramp vector}
}
\description{
Inner color function setup
}
