% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca_estimation_and_visualization.R
\name{data_projection}
\alias{data_projection}
\title{Get points for objects in projection space}
\usage{
data_projection(scp_output)
}
\arguments{
\item{scp_output}{pectral_cluster_process function output}
}
\value{
\item{test_df}{data frame of training points and probability weights}
\item{train_df}{data frame of test points}
}
\description{
Get points for objects in projection space
}
