% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visual_functions.R
\name{get_kde_contour_path}
\alias{get_kde_contour_path}
\title{Create level set for kde}
\usage{
get_kde_contour_path(kde.obj, level = "5\%")
}
\arguments{
\item{kde.obj}{kde object}

\item{level}{string of percentage level outside the contour}
}
\value{
data frame of countour (detail determined by detail level in kde.obj)
}
\description{
Returns the level set for a given kde
}
