% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visual_functions.R
\name{ggvis_kde_contour}
\alias{ggvis_kde_contour}
\title{Create ggplot of contour}
\usage{
ggvis_kde_contour(level_contour_list, base_graph = NULL, zoom = 4,
  color = "pink", ...)
}
\arguments{
\item{level_contour_list}{list of  data frame of level contours}

\item{base_graph}{ggplot object for base graph 
(created from data_out otherwise)}

\item{zoom}{map zoom for ggmap}

\item{color}{color of band}

\item{...}{interior ggpath parameters}
}
\value{
ggplot object of contour and data points.
}
\description{
Create ggplot of contour
}
