% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca_estimation_and_visualization.R
\name{ggvis_paths_sca_weight}
\alias{ggvis_paths_sca_weight}
\title{Visualize the TC paths with weights colored}
\usage{
ggvis_paths_sca_weight(data_out, zoom = 4, test_color_power = 1/3,
  test_color_low = "white", test_color_high = "red", base_graph = NULL)
}
\arguments{
\item{data_out}{data frame with correct path information same as outputed 
from data_plot_sc_paths functions}

\item{zoom}{map zoom for ggmap}

\item{test_color_power}{power to raise the probability weights 
(then use linear scaling for colors)}

\item{test_color_low}{lower color value for color range,}

\item{test_color_high}{higher color value for color range,}

\item{base_graph}{ggplot object for base graph 
(created from data_out otherwise)}
}
\value{
ggmap based map object
}
\description{
Note / TODO: 
1. Function currently uses geom_path - assumed euclidean space for map :/
}
