% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca_estimation_and_visualization.R
\name{ggvis_projection}
\alias{ggvis_projection}
\title{Creates ggplot of projections}
\usage{
ggvis_projection(scp_output, train_alpha = 0.3, test_color_power = 1/3,
  test_color_low = "white", test_color_high = "red")
}
\arguments{
\item{scp_output}{spectral_cluster_process function output}

\item{train_alpha}{opacity level for black training points}

\item{test_color_power}{power transformation (x^ test_color_power) of 
probability values for test points color}

\item{test_color_low}{lower color for range of colors on test points prob}

\item{test_color_high}{high color for range of colors on test points prob}
}
\value{
ggplot scatter plot of training and colored test points
}
\description{
Creates ggplot of projections
}
\details{
ggplot experts are encouraged to use the output of data_projection 
function applied to scp_output instead of this wrapper
}
