% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca_estimation_and_visualization.R
\name{kernel_estimate}
\alias{kernel_estimate}
\title{KNN density estimate for each point}
\usage{
kernel_estimate(train, test, k)
}
\arguments{
\item{train}{Training points}

\item{test}{Testing points}

\item{k}{number of neighbors that matter}
}
\value{
p density estimate for test observations
}
\description{
KNN in Euclidean Distance
}
